stickied comments will no longer be used
I made a video but it was taken up entirely by a post looking like this: https://imgur.com/a/1fwT94c
This isn't a big deal because these serious tags are uncommon, but still a small annoyance.
So I fixed it!
In threads that are marked 'serious', there is an automatic stickied mod post. ex. (shorturl.at/I1346). My changes will skip any automatically stickied comments, so they won't make it into the final Video.