Better anti NSFW
Instead of looking at the screen to look if it's nsfw it checks it earlier so it doesn't have to create all the mp3's and open the browser to do it. Also it fixes some errors because of it cause playwright was being buggy with it for some reason.