Allow SUBREDDIT to be blank
This feature was in an old version but was removed in a recent PR. It allows the env variable SUBREDDIT to be blank and the user input to be blank which will default to AskReddit instead of an error. It also automatically filters out any special characters (in case someone types r/) from the subreddit using re.