Allow TTsChoice and ALLOW_NSFW to be blank
All these changes to the .env file make it annoying to keep updating. This allows the TTsChoice and ALLOW_NSFW variables to be left blank in the event that someone doesn't update the .env file. Also, videos.json in the .gitignore was not seen by GH Desktop, and it kept wanting to push that too. I made a change in the .gitignore.