Edited main.py and CONTRIBUTING.md
Added a line to code to the shutdown function to inform the user of exiting the program upon calling the function.

And,

Also wrote a section of CONTRIBUTING.md to help beginner developers to setup their fork of the repo as many beginners to github might not know how to setup this up.