Adding prawcore to the requirements
Multiple people on the discord ran into the issue that an error message would come up saying "ModuleNotFoundError: No module names 'prawcore'". I was told praw itself is supposed to install it but apparently it does not? I cannot verify if the people actually followed the steps correctly. Json told me to make a PR for it :) I took the value for the requirements.txt from here: https://pypi.org/project/prawcore/