Added multi-python check, colors, and run menu
Added the ability to run the script with multiple versions of python installed and selected the proper version for the virtual environment. Added pretty colors. create a menu out of the run.bat that allows you to install the The prerequisites run the bot or open a command prompt within the installed environment.